IB/requirements.txt
fastapi
uvicorn
structlog
PyJWT
duckduckgo-search
anthropic>=0.3.0
trubrics>=1.4.3



#imagebind install
torch==1.13.0
torchvision==0.14.0
torchaudio==0.13.0
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
timm==0.6.7
ftfy
regex
einops
fvcore
eva-decord==0.6.1
iopath
numpy>=1.19
matplotlib
types-regex
mayavi
cartopy
pandas
opencv-python==4.8.0.76
moviepy
python-dotenv
elasticsearch
python-multipart


#STT
google-cloud-speech==2.21.0
librosa
